requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
openai==0.27.6
python-dotenv==1.0.0
